Disable validation when unmarshalling only part of the XML file
Validation is disabled when unmarshalling from a String. This type of
unmarshalling is commonly used to unmarshal only part of the XML file.
However, validation would always fail because only the qcML-element is
allowed as root by the XML schema. Therefore, only perform validation
when unmarshalling from a full file, and temporary disable validation
when unmarshalling from a String.